Create mac-scroll class that fixes disappearing scrollbar problem
Moved Mac scrollbar fix into main.css, attached to `mac-scroll` class
Applied to TOC on details page
Moved freeze-main-window-when-scrolling-popover functionality into
BaseView (requires call to `super.onRender()` in derived class to
activate)
