+++ src/app/components/single-select/single-select.hbs
@@ -1,2 +1,2 @@
 <div class="selected-button has-dropdown"></div>
-<div class="option-list hidden"></div>
+<div class="option-list mac-scroll hidden"></div>
