+++ src/app/components/tag-multi-select/tag-multi-select.hbs
@@ -1,3 +1,3 @@
 <div class="tag-list has-dropdown"></div>
-<div class="option-list hidden"></div>
+<div class="option-list mac-scroll hidden"></div>
 <span class="required"></span>
