+++ src/app/pages/details/details.hbs
@@ -9,7 +9,7 @@
                     <h1 class="title"></h1>
                     <div class="tabs">
                         <a class="table-of-contents-link" class="tab"><span>Table of Contents</span></a>
-                        <div class="table-of-contents hidden">
+                        <div class="table-of-contents mac-scroll hidden">
                             <div class="box">
                                 <div class="toc-remover"></div>
                                 <ol></ol>
