+++ src/styles/components/option-list.scss
@@ -20,25 +20,6 @@ $faint-gray: rgba(color(gray), 0.33);
     border-left-color: rgba(color(gray), 0.33);
     box-shadow: 0 0.2rem 0.3rem rgba(color(neutral-lighter), 0.8);
 
-    &::-webkit-scrollbar {
-        -webkit-appearance: none;
-    }
-
-    &::-webkit-scrollbar:vertical {
-        width: 1.1rem;
-    }
-
-    &::-webkit-scrollbar-thumb {
-        border-radius: 0.8rem;
-        border: 0.2rem solid white; /* should match background, can't be transparent */
-        background-color: rgba(0, 0, 0, 0.5);
-    }
-
-    &::-webkit-scrollbar-track {
-        background-color: #fff;
-        border-radius: 0.8rem;
-    }
-
     > .option {
         background-color: color(neutral-lightest);
         font-weight: 600;
